refactor: split user permission tasks
