+++ handlers/blogs/blogsUserPermissionsPage.go
@@ -16,100 +16,7 @@ import (
 	"strings"
 
 	"github.com/arran4/goa4web/handlers"
-	"github.com/arran4/goa4web/internal/eventbus"
-	notif "github.com/arran4/goa4web/internal/notifications"
-	"github.com/arran4/goa4web/internal/tasks"
 )
-
-// UserAllowTask grants a user a permission.
-type UserAllowTask struct{ tasks.TaskString }
-
-var userAllowTask = &UserAllowTask{TaskString: TaskUserAllow}
-
-var _ tasks.Task = (*UserAllowTask)(nil)
-var _ notif.AdminEmailTemplateProvider = (*UserAllowTask)(nil)
-var _ notif.TargetUsersNotificationProvider = (*UserAllowTask)(nil)
-
-func (UserAllowTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationBlogUserAllowEmail")
-}
-
-func (UserAllowTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationBlogUserAllowEmail")
-	return &v
-}
-
-func (UserAllowTask) Action(w http.ResponseWriter, r *http.Request) any {
-	UsersPermissionsPermissionUserAllowPage(w, r)
-	return nil
-}
-
-// UserDisallowTask removes a user's permission.
-type UserDisallowTask struct{ tasks.TaskString }
-
-var userDisallowTask = &UserDisallowTask{TaskString: TaskUserDisallow}
-
-var _ tasks.Task = (*UserDisallowTask)(nil)
-var _ notif.AdminEmailTemplateProvider = (*UserDisallowTask)(nil)
-var _ notif.TargetUsersNotificationProvider = (*UserDisallowTask)(nil)
-
-func (UserDisallowTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationBlogUserDisallowEmail")
-}
-
-func (UserDisallowTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationBlogUserDisallowEmail")
-	return &v
-}
-
-func (UserDisallowTask) Action(w http.ResponseWriter, r *http.Request) any {
-	UsersPermissionsDisallowPage(w, r)
-	return nil
-}
-
-// UsersAllowTask grants multiple users permissions.
-type UsersAllowTask struct{ tasks.TaskString }
-
-var usersAllowTask = &UsersAllowTask{TaskString: TaskUsersAllow}
-
-var _ tasks.Task = (*UsersAllowTask)(nil)
-var _ notif.AdminEmailTemplateProvider = (*UsersAllowTask)(nil)
-
-func (UsersAllowTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationBlogUsersAllowEmail")
-}
-
-func (UsersAllowTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationBlogUsersAllowEmail")
-	return &v
-}
-
-func (UsersAllowTask) Action(w http.ResponseWriter, r *http.Request) any {
-	UsersPermissionsBulkAllowPage(w, r)
-	return nil
-}
-
-// UsersDisallowTask removes permissions from multiple users.
-type UsersDisallowTask struct{ tasks.TaskString }
-
-var usersDisallowTask = &UsersDisallowTask{TaskString: TaskUsersDisallow}
-
-var _ tasks.Task = (*UsersDisallowTask)(nil)
-var _ notif.AdminEmailTemplateProvider = (*UsersDisallowTask)(nil)
-
-func (UsersDisallowTask) AdminEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("adminNotificationBlogUsersDisallowEmail")
-}
-
-func (UsersDisallowTask) AdminInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("adminNotificationBlogUsersDisallowEmail")
-	return &v
-}
-
-func (UsersDisallowTask) Action(w http.ResponseWriter, r *http.Request) any {
-	UsersPermissionsBulkDisallowPage(w, r)
-	return nil
-}
 
 func GetPermissionsByUserIdAndSectionBlogsPage(w http.ResponseWriter, r *http.Request) {
 	cd := r.Context().Value(consts.KeyCoreData).(*common.CoreData)
@@ -305,42 +212,4 @@ func roleInfoByPermID(ctx context.Context, q *db.Queries, id int32) (int32, stri
 		}
 	}
 	return 0, "", "", sql.ErrNoRows
-}
-
-func (UserAllowTask) TargetUserIDs(evt eventbus.TaskEvent) ([]int32, error) {
-	if id, ok := evt.Data["targetUserID"].(int32); ok {
-		return []int32{id}, nil
-	}
-	if id, ok := evt.Data["targetUserID"].(int); ok {
-		return []int32{int32(id)}, nil
-	}
-	return nil, fmt.Errorf("target user id not provided")
-}
-
-func (UserAllowTask) TargetEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("setUserRoleEmail")
-}
-
-func (UserAllowTask) TargetInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("set_user_role")
-	return &v
-}
-
-func (UserDisallowTask) TargetUserIDs(evt eventbus.TaskEvent) ([]int32, error) {
-	if id, ok := evt.Data["targetUserID"].(int32); ok {
-		return []int32{id}, nil
-	}
-	if id, ok := evt.Data["targetUserID"].(int); ok {
-		return []int32{int32(id)}, nil
-	}
-	return nil, fmt.Errorf("target user id not provided")
-}
-
-func (UserDisallowTask) TargetEmailTemplate() *notif.EmailTemplates {
-	return notif.NewEmailTemplates("deleteUserRoleEmail")
-}
-
-func (UserDisallowTask) TargetInternalNotificationTemplate() *string {
-	v := notif.NotificationTemplateFilenameGenerator("delete_user_role")
-	return &v
 }
